Updated the blog controller

+++ src/app/modules/blog/blog.service.ts
@@ -1,10 +1,12 @@
-import httpStatus, { REQUEST_URI_TOO_LONG } from 'http-status';
+
+import httpStatus from 'http-status';
 import AppError from '../../errors/AppError';
 import { TBlog } from './blog.interface';
 import { BlogModel } from './blog.model';
 
 const createBlogIntoDB = async (payload: TBlog) => {
   const result = await BlogModel.create(payload);
+  
   return result;
 };
 
@@ -13,11 +15,19 @@ const getBlogFromDb = async (query: Partial<TBlog>) => {
   return result;
 };
 const getSingleBlogFromDb = async(id:string)=>{
-    const result = await BlogModel.find({_id:id})
-    return result
+    const result = await BlogModel.findOne({_id:id})
+    if(!result){
+      throw new AppError(httpStatus.NOT_FOUND,"Coludn't found the item")
+    }
+    else{
+      return result
+    }
+    
 }
 
 const updateBlogFromDb = async (id: string, payload: Partial<TBlog>) => {
+  console.log(id)
+  console.log(payload)
   const result = await BlogModel.findOneAndUpdate({ _id: id }, payload, {
     new: true,
   });
@@ -25,6 +35,10 @@ const updateBlogFromDb = async (id: string, payload: Partial<TBlog>) => {
 };
 
 const deleteBlogFromDb = async (id: string) => {
+  const findData = await BlogModel.findById(id)
+  if(!findData){
+    throw new AppError(httpStatus.NOT_FOUND,"Coludn't found the data")
+  }
   const result = await BlogModel.deleteOne({ _id: id });
   return result;
 };
